requirement.txt
azure-eventhub
azure-eventhub-checkpointstoreblob-aio
requests